pylint fixes
